Construct workflows collection url from filtering query parameters
Front end support for filtering workflows, #12. Using search and radio button UI components to update a filter variable in the view with values for a search term, input type, and/or dwc class.

The view constructs a new collection and fetches from the backend by including the filter criteria in the url query parameters.